nas-backup-system/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
psutil==5.9.6
python-dotenv==1.0.0
